[v1] Get everything ready for the first release
This is just the previous version, packaged as a separate module
