--- src/constants.js
@@ -1,2 +1,0 @@
-
-export const BROADCAST_PREFIX = 'app-broadcast:';
